Refactor upsertJob to use new class
